+++ plugins/sdk/src/index.ts
@@ -1,2 +1,4 @@
 export * from './lib/sdk.module';
-export * from './lib/navigation.service';
+
+export * from './lib/components';
+export * from './lib/services';